Add PHPUnit tests for the minification failure case in bug 32548.
This will trigger 2 test failures, where an exponent in a JS numeric literal gets split over line breaks at the '-' or '+', causing a parse error in the resulting output.
A number with the same string length but without using + or - in the exponent passes through fine, indicating that it's the -/+ that's getting misinterpreted.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@103846 dd0e9695-b195-4be7-bd10-2dea1a65a6b6