* (bug 16434) Fix paging on Special:Prefixindex when prefix is empty (whole namespace).
Tweaked input handling to make sure the 'from' => 'prefix' mapping doesn't happen if we have passed a namespace parameter (indicating we're working within the form). Tweaked the form paging links to include the NS even if it's 0. :)
Originally wanted to just check for presence of 'prefix' however wfArrayToCgi() strips empty values, and I'm not comfortable tweaking that without a lot more testing.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@104765 dd0e9695-b195-4be7-bd10-2dea1a65a6b6